Rename package names of logic, logicCommands & viewCommands
<old package name> -> <new package name>
logic -> internal
commands -> command
logicCommand -> logic
viewCommand -> view
