+++ src/main/java/vimification/Gui.java
@@ -13,8 +13,8 @@ import vimification.commons.core.LogsCenter;
 import vimification.commons.exceptions.DataConversionException;
 import vimification.commons.util.ConfigUtil;
 import vimification.commons.util.StringUtil;
-import vimification.logic.Logic;
-import vimification.logic.LogicManager;
+import vimification.internal.Logic;
+import vimification.internal.LogicManager;
 import vimification.model.LogicTaskList;
 import vimification.model.ReadOnlyUserPrefs;
 import vimification.model.UserPrefs;
