+++ src/main/java/vimification/internal/command/CommandException.java
@@ -1,4 +1,4 @@
-package vimification.logic.commands;
+package vimification.internal.commands;
 
 /**
  * Represents an error which occurs during execution of a {@link Command}.
