+++ src/main/java/vimification/internal/command/logic/CreateCommand.java
@@ -1,9 +1,9 @@
-package vimification.logic.commands.logicCommands;
+package vimification.internal.commands.logic;
 
 import static java.util.Objects.requireNonNull;
 
-import vimification.logic.commands.CommandException;
-import vimification.logic.commands.CommandResult;
+import vimification.internal.commands.CommandException;
+import vimification.internal.commands.CommandResult;
 import vimification.model.LogicTaskList;
 import vimification.model.task.Task;
