+++ src/main/java/vimification/internal/command/logic/UndoableLogicCommand.java
@@ -1,7 +1,7 @@
-package vimification.logic.commands.logicCommands;
+package vimification.internal.commands.logic;
 
-import vimification.logic.commands.CommandException;
-import vimification.logic.commands.CommandResult;
+import vimification.internal.commands.CommandException;
+import vimification.internal.commands.CommandResult;
 import vimification.model.LogicTaskList;
 
 public abstract class UndoableLogicCommand extends LogicCommand {
