+++ src/main/java/vimification/internal/command/oldcode/FindCommand
