+++ src/main/java/vimification/internal/parser/Pair.java
@@ -1,4 +1,4 @@
-package vimification.logic.parser;
+package vimification.internal.parser;
 
 import java.util.function.Function;
