+++ src/main/java/vimification/internal/parser/Prefix.java
@@ -1,4 +1,4 @@
-package vimification.logic.parser;
+package vimification.internal.parser;
 
 /**
  * A prefix that marks the beginning of an argument in an arguments string. E.g. 't/' in 'add James
