+++ src/main/java/vimification/internal/parser/VimificationParser.java
@@ -1,9 +1,9 @@
-package vimification.logic.parser;
+package vimification.internal.parser;
 
 import java.util.logging.Logger;
 
 import vimification.commons.core.LogsCenter;
-import vimification.logic.commands.logicCommands.LogicCommand;
+import vimification.internal.commands.logic.LogicCommand;
 
 public class VimificationParser {
