+++ src/test/java/vimification/logic/parser/CreateCommandParserTest.java
@@ -5,7 +5,9 @@ import static org.junit.jupiter.api.Assertions.assertTrue;
 
 import org.junit.jupiter.api.Test;
 
-import vimification.logic.commands.logicCommands.CreateCommand;
+import vimification.internal.commands.logic.CreateCommand;
+import vimification.internal.parser.CreateCommandParser;
+import vimification.internal.parser.ParserException;
 
 public class CreateCommandParserTest {
