+++ src/test/java/vimification/logic/parser/VimificationParserTest.java
@@ -6,7 +6,9 @@ import static org.junit.jupiter.api.Assertions.assertTrue;
 
 import org.junit.jupiter.api.Test;
 
-import vimification.logic.commands.logicCommands.LogicCommand;
+import vimification.internal.commands.logic.LogicCommand;
+import vimification.internal.parser.ParserException;
+import vimification.internal.parser.VimificationParser;
 
 public class VimificationParserTest {
